Some minor renames to modules
