+++ .vscode/settings.json
@@ -28,6 +28,7 @@
         "Invokable",
         "keyvaluestore",
         "ledgertransaction",
+        "Mergeable",
         "notarizedtransaction",
         "pdec",
         "Peekable",
@@ -52,6 +53,7 @@
         "transactiontracker",
         "txid",
         "undeposited",
+        "Unmergeable",
         "unpreparable",
         "unregistration",
         "Unstake",
